worker/pyproject.toml
[project]
name = "worker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beanie>=2.0.1",
    "kubernetes>=34.1.0",
    "motor>=3.7.1",
    "python-dotenv>=1.2.1",
    "redis>=7.1.0",
]
